+++ src/react/index.ts
@@ -3,7 +3,9 @@ export { useStorageState } from './useStorageState';
 export { useStorageReducer } from './useStorageReducer';
 export { useStorageSync } from './useStorageSync';
 export { useStorageEvent } from './useStorageEvent';
+export { useStorageList } from './useStorageList';
 export { StorageProvider, useStorageContext } from './StorageContext';
 
 export type { UseStorageStateResult } from './useStorageState';
+export type { UseStorageListActions, UseStorageListResult } from './useStorageList';
 export type { StorageProviderProps } from './StorageContext';